solved the issue with the mysterious points being affected...
it was because pressure is not calculated at some points and pressure is
calculated at other points. as a result the wall points have this bogus
original pressure that causes trouble. i got rid of the envifwall and
positionifwall callsl in prop, which speeds things up and now rely on
the fall-through to initial grid in the prop get maybe thing.